requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.7
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
discord-py==2.4.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
yarl==1.18.0
    # via aiohttp
